Move redundant_allocation to its own module
